pyproject.toml
[project]
name = "livekit-agent"
version = "0.1.0"
description = "LiveKit Voice Agent with Agno Integration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=2.4.7",
    "livekit-agents[deepgram,openai,silero]~=1.3",
    "python-dotenv>=1.0.0",
]

